Add Invoices logic to nubefact

+++ products/models.py
@@ -25,7 +25,7 @@ class ProductModel(models.Model):
     id = models.AutoField(primary_key=True)
     name = models.CharField(max_length=100)
     description = models.TextField()
-    price = models.DecimalField(max_digits=10, decimal_places=2)
+    price = models.FloatField()
     image = CloudinaryField('image')
     color = ArrayField(models.CharField(max_length=10, null=True), default=list, blank=True) # ['red', 'blue', 'green'] array de colores
     stock = models.IntegerField()
@@ -42,7 +42,7 @@ class ProductModel(models.Model):
 
 class SaleModel(models.Model):
     id = models.AutoField(primary_key=True)
-    total_price = models.DecimalField(max_digits=10, decimal_places=2)
+    total_price = models.FloatField()
     user_id = models.ForeignKey(MyUser, on_delete=models.CASCADE)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
@@ -56,8 +56,8 @@ class SaleModel(models.Model):
 class SaleDetailModel(models.Model):
     id = models.AutoField(primary_key=True)
     quantity = models.IntegerField()
-    price = models.DecimalField(max_digits=20, decimal_places=2)
-    subtotal = models.DecimalField(max_digits=10, decimal_places=2)
+    price = models.FloatField()
+    subtotal = models.FloatField()
     product_id = models.ForeignKey(ProductModel, on_delete=models.CASCADE)
     sale_id = models.ForeignKey(SaleModel, on_delete=models.CASCADE, related_name='sale_details')
